requirements.txt
requests
lxml==4.9.3
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
